Spring controller tests
