+++ integration-server/src/test/java/com/canoo/dolphin/integration/server/action/ActionTestControllerTest.java
@@ -3,9 +3,9 @@ package com.canoo.dolphin.integration.server.action;
 import com.canoo.dolphin.client.Param;
 import com.canoo.dolphin.integration.action.ActionTestBean;
 import com.canoo.dolphin.integration.server.TestConfiguration;
-import com.canoo.dolphin.test.ControllerTestException;
-import com.canoo.dolphin.test.ControllerUnderTest;
-import com.canoo.dolphin.test.SpringTestNGControllerTest;
+import com.canoo.platform.spring.test.ControllerTestException;
+import com.canoo.platform.spring.test.ControllerUnderTest;
+import com.canoo.platform.spring.test.SpringTestNGControllerTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
