+++ platform-examples/todo-example/todo-server-spring/src/test/java/com/canoo/dolphin/todo/server/ToDoControllerTests.java
@@ -16,8 +16,8 @@
 package com.canoo.dolphin.todo.server;
 
 import com.canoo.dolphin.client.Param;
-import com.canoo.dolphin.test.ControllerUnderTest;
-import com.canoo.dolphin.test.SpringTestNGControllerTest;
+import com.canoo.platform.spring.test.ControllerUnderTest;
+import com.canoo.platform.spring.test.SpringTestNGControllerTest;
 import com.canoo.dolphin.todo.pm.ToDoList;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.testng.Assert;
