+++ platform-extras/dolphin-platform-spring-controller-tester/src/main/java/com/canoo/platform/spring/test/ControllerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.test;
+package com.canoo.platform.spring.test;
 
 import com.canoo.platform.server.DolphinController;
