+++ platform-extras/dolphin-platform-spring-controller-tester/src/main/java/com/canoo/platform/spring/test/ControllerTestException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.test;
+package com.canoo.platform.spring.test;
 
 /**
  * Exception that is used when testing controllers.
